+++ innovation.cabal
@@ -29,10 +29,7 @@ library
                      , Game.MetaGame.Play
                      , Game.MetaGame.Types
                      , Game.MetaGame.Types.Core
-                     , Game.MetaGame.Types.Board
                      , Game.MetaGame.Types.GameState
-                     , Game.MetaGame.Types.Game
-                     , Game.MetaGame.Types.UserInput
                      , Game.MetaGame.Types.GameRules
 
                      , Game.Innovation
