+++ lib/Game/MetaGame.hs
@@ -3,6 +3,9 @@ module Game.MetaGame
        ) where
 
 import           Game.MetaGame.Types as X
+  hiding ( InnerMoveType, InnerMoveResult, runInnerMoveType
+         , OuterMoveResult, liftFromInner, runOuterMoveType
+         )
 import           Game.MetaGame.Helper as X
 import           Game.MetaGame.Ask as X
 import           Game.MetaGame.Play as X
